Tweak hyphen as a togglable option after all
In 1818bf0c43b8aad0f302b3e4d97324cfbf8b00c3 I wondered why hyphenation was set
to be a toggle able feature when it opens a non-modal dialog.

Looking closer, it *is* a toggable feature on/off for draw/impress, but this
isn't the case for writer. Reverted it for calc too, although apparently
not toggable there.

Change-Id: If02568d58253c380e6d668a728f9c3ae6acf1128