More libraries
Now the DocumentLoader test app seems to work like it does with all
the separate DSOs, knock on wood. The only "attempting to load unknown
library" message I see in logcat is for libspelllo.a, and I
deliberately don't have a mapping for that. A viewer application
presumably has no use for spell checking. (I do link in the
hyphenation stuff, that affects rendering, and we want the rendering
to match LO on the desktop.)

Change-Id: I1ad9eedf352eb0af1187bdfdc71f8f1754912600