Adapt to disable-dynloading style
No need to load any native libraries here any more.

For now, using the same native-code.cxx as over in ../DocumentLoader
works. In other words, we end up with the same code in
liblo-native-code.so. This will obviously change if/when the UNO APIs
that are called start to differ.

Change-Id: Ia47ce833c5a290a93986005f92e981c89f63bf71