sd: fix crash in character format dialog:
When text with different fonts is selected in Draw or Impress, the
character format dialog crashes, because (among other things) the
selected text is not returned by SfxViewShell::GetSelectionText().
Crash was observed both with GraphicViewShellBase and
ImpressViewShellBase as current shell.
Override that method in sd::ViewShellBase to forward to a main shell
that happens to be a DrawViewShell which is not actually a SfxViewShell
but happens to implement GetSelectionText() for mysterious reasons.