make configure remember C/CXX/OBJC/OBJCXX/LDFLAGS if explicitly passed to it
If not, the build system selects the right flags depending on --enable-debug
etc. , but e.g. packagers may want to do builds with e.g. both -O and -g.
This is also reasonably close to what autotools do, except that there configure
forces -g -O2 if nothing is explicitly given.

Change-Id: I1889569177ad3cbd6decdc7e9986f67ef5d4456a