sw: run export tests twice
Export tests are in fact import-export-import tests. Given that we want
to test the export filter here, it's handy to run the tests twice, first
before the export (to detect if already the import filter fails) and
also after the second import, to do the real test. So this should make
it easier to detect when you have an export failure, but in fact the bug
is in the import side.

Change-Id: I26e3445f079a7fbd4e037218f7094a84e730b746
Requested-by: Pierre-Eric Pelloux-Prayer <pierre-eric@lanedo.com>
Agreed-by: Lubos Lunak <l.lunak@suse.cz>