Resolves: fdo#50285 merging props unsafe with intermediate charstyle prop
Its a nice idea to merge properties of the same type and value if one starts
and the same place the last one ended. But character styles are properties as
well, so if we have character-style+superscript on range a, and
character-style+superscript on range b and merge these so that we end up as...

character-style on range a, superscript on range a+b, and character-style on
range b then that clearly gives the wrong result if applied in that order.

So its only safe to merge if there are no intermediate properties that can
affect the merge candidates.

A regression from 9d1b5529366da2e98518976a2dd2b4dac52cca16

Change-Id: I541563d11265426736b840de068922eef8d45573