fdo#55290: use the right native library name
It turns out that the native library cannot be renamed after creation.
The right name must be used from the beginning, otherwise
publishing is failing with the follow error (gacutil -i):

"Failure adding assembly to the cache: Invalid file or assembly name.
The name of the file must be the name of the assembly plus .dll or .exe."

To rectify that create the native lib as cli_cppuhelper.dll,
rename the signed assembly to assembly/cli_cppuhelper.dll and
teach scp2 to pick up the right one: assembly/cli_cppuhelper.dll
(and not cli_cppuhelper.dll).

Change-Id: I2073b617cd440865ae4ab838bb801329f2b07194