Don't let NSS use the MSVC debugging runtime in an --enable-debug build
In an MSVC build, not exporting BUILD_OPT to the Mozilla build
machinery causes the produced DLLs to use the debug CRT. The exact
mechanism is a bit of a mystery, and I didn't feel like spending too
much time trying to understand it.

Using the debug CRT is confusing and wrong. Nothing in LO otherwise
uses it. It also makes testing a build much harder for me at least, as
I do that in a fairly pristine virtual machine with no MSVC debugging
runtime available. (The normal CRT is bundled in the LO installer.)

Change-Id: I27f774d92a3986d40162c870202bcdddd94aa7c6