Revert "fdo#38878 upgrade xpdf to poppler"
Also revert my two minor follow-up popper commits.

Poppler doesn't build on Mac OS X. It uses fontconfig, and we don't
depend on that on the Mac.

There is a patch from 2009
http://lists.freedesktop.org/archives/poppler/2009-July/004932.html
(actual patch at
http://lists.freedesktop.org/archives/poppler/attachments/20090712/f3b154d0/attachment-0001.obj
) that makes poppler not use fontconfig on the Mac, but it was never
upstreamed.

And that patch uses the obsolete ATSUI API, which we want to stop
using in LO code, not introduce new uses of.

This reverts commit 861c61f1636741b5c0f4b2b6e55007ec92a91bff.
This reverts commit d811cd5c713547cdf5ec3aac2a1189d94185f98a.
This reverts commit b994b08bbfe929db1af3d7a23e9206093bdb9563.