n#782833: Rotated text in emf images looks thicker.
Any rotated text in an emf image (image size != original size),
has thicker fonts.
This is kind of a regression caused by commit
8bf84a2b0ea3f0e79677ad66cd5204c50110cd6d
So, partially reverted it.
The code was kind of a workaround/copy-paste, so I am assuming
it is safe to revert.

Side-effects: The dotted borders (in calc) look a
little lighter and not so very visible in the page-preview.