make the word count dialog the right size with disabled CJK word counting
we want to hide the CJK word count features unless CJK UI is enabled, or if
there are CJK chars in the word count. We want the dialog to fit to the best
size, so force the dialog to resize after it's visible (like a vclexpander)
if in non-CJK mode and some CJK chars appear in the document.

Change-Id: I328dcf2b111d5c5f527757ca11695ce4e9c43353