Add experimental --disable-export configure option
Even for an Android (or iOS, presumably) app with just viewer
functionality, some export code gets included in the single .so or
executable.

For now, when we know that we are producing only viewer apps for
Android and iOS anyway, it might be a good idea to be able to
carefully bypass some particularly pathological places in the code
that generate large amounts of code and/or data that is useless in a
viewer app.

Change-Id: I59b97024ff2855e7bbe53d6ec8782797705f61d8