gbuild: UnoApiTarget: fix spurious rebuilds:
Since commit 14207317b661d6446268b3fedacfa60ab4f7b027 there are spurious
rebuilds when a IDL file is rebuilt, because the rule for the idlc
invocation was changed to rebuild all IDL files, but the dummy rule for
.urd files only touched that when its own IDL file changes; this means
that the header target is not seen as outdated in this make run, but it
will be outdated in the next make run because then the .urd file
timestamp is checked and is newer.
The dummy rule for .urd files must touch the .urd file if and only if
the rule for the .done file re-builds that .urd file.

Change-Id: I37938aef0621c7d46809e02a06d22248de28271b