We need to tell the MSVS 2010 build to use Windows7.1SDK
At least for me it wouldn't build otherwise. But yeah, what it
somebody uses MSVS 2010 with another SDK? It seems that the solution
only offers the SDK 7.1 as an alternative?

The default was v100, whatever that measn. Could it be that my MSVS
2010 installation is borked? Or that I did not have to install a
bundled SDK with it, because I already had a separate 7.1 SDK?

Also simplify a bit, no need to $(filter) on VCVER inside ifeqs that
already check the very same VCVER.

Change-Id: Ifef98c9466fc24db27d9e38c6878c77adfb4ed75