fdo#55289: fix crash in SwXShape::setPropertyValue:
In the libreoffice-3-6 branch the docx import crashes here when setting
the anchor position of a shape, because the previous setting of the
AnchorType was not done properly: the position of the anchor in the
SwFmt did not match the actual position of the text attribute, so we get
these assertions: "Missing FlyInCnt-Hint." and crash.

This cannot be reproduced with the docx import on master, because on
master a different intermediate AnchorType is set, and transitioning
from that happens not to cause the crash.  But it can be reproduced with
a unit test.

The regression was introduced in libreoffice-3-6 with
backport commit 9e15a8c87221aeda2e37ed4fab4138db8e2bf940.

Change-Id: I394643e11862543177ba57958a26cd8ce06dc09c