use config_xxx.h files instead of -DHAVE_XXX on the command line
As the latter does not quite scale, and also source files depending
on the setting/feature did not rebuild in case of a change.

There are intentionally more config_xxx.h files (so autoheader
from autotools is not used), so that a setting change does not force
automatically a rebuild of everything. Running configure does not
touch those config_xxx.h files that have not changed.

There's config/README with a howto.

Change-Id: I6d6e82050ff8faf5364ed246848b8401aca3e0e5